refactor(tokens): move primitive platform config to scripts/config
Style Dictionary platform configs are data (source globs + platform
definitions), not reusable logic, so scripts/lib is the wrong home once more
tiers (semantic, component) land alongside it. Move
scripts/lib/primitive-platforms.mjs to scripts/config/primitive.config.mjs
and give it a proper @file header instead of a run-on top comment.

scripts/lib now holds only genuinely reusable utilities (transforms.mjs,
token-tooling.mjs). Future semantic/component tier configs get an obvious,
pre-named home under scripts/config/.
