+++ packages/ontario-design-system-design-tokens/test/primitive-platforms.test.mjs
@@ -1,6 +1,6 @@
 // Tests for the primitive layer's Style Dictionary output platforms (DS-2691):
 // the `primitives.*` CSS/SCSS/JS/TS entry points built from
-// `scripts/lib/primitive-platforms.mjs`, registered via the transforms in
+// `scripts/config/primitive.config.mjs`, registered via the transforms in
 // `scripts/lib/transforms.mjs`.
 import { describe, it, expect, beforeAll } from 'vitest';
 import { readFileSync } from 'node:fs';
@@ -9,7 +9,7 @@ import { fileURLToPath } from 'node:url';
 import StyleDictionary from 'style-dictionary';
 
 import { primitiveTransforms } from '../scripts/lib/transforms.mjs';
-import { primitivePlatformsConfig } from '../scripts/lib/primitive-platforms.mjs';
+import { primitivePlatformsConfig } from '../scripts/config/primitive.config.mjs';
 
 const packageRoot = path.dirname(path.dirname(fileURLToPath(import.meta.url)));
